user-workspace/swarm/agents/MEVAgent.py
import asyncio
from typing import Dict, List, Optional
import aiohttp
from decimal import Decimal
import json
from web3 import Web3

class MEVAgent:
    """Agent that captures MEV opportunities through sandwich attacks and arbitrage"""
    
    def __init__(self, wallet_manager, initial_capital: float = 2000.0):
        self.wallet_manager = wallet_manager
        self.capital = initial_capital
        self.min_profit = 0.3  # 0.3% minimum profit threshold
        self.w3 = Web3(Web3.HTTPProvider('https://polygon-rpc.com'))
        self.monitored_pools = [
            "0xa5E0829CaCEd8fFDD4De3c43696c57F7D7A678ff",  # QuickSwap
            "0x1b02dA8Cb0d097eB8D57A175b88c7D8b47997506"   # SushiSwap
        ]
        
    async def monitor_mempool(self) -> List[Dict]:
        """Monitor mempool for potential MEV opportunities"""
        opportunities = []
        
        try:
            # Get pending transactions
            pending_txs = self.w3.eth.get_pending_transactions()
            
            for tx in pending_txs:
                # Look for large swaps in DEX pools
                if tx['to'] in self.monitored_pools and tx['value'] > self.w3.toWei(1, 'ether'):
                    # Analyze transaction for frontrunning opportunity
                    opportunity = await self.analyze_transaction(tx)
                    if opportunity:
                        opportunities.append(opportunity)
                        
        except Exception as e:
            print(f"Mempool monitoring error: {e}")
            
        return opportunities
        
    async def analyze_transaction(self, tx: Dict) -> Optional[Dict]:
        """Analyze transaction for MEV opportunity"""
        try:
            # Decode transaction input
            decoded_input = self.decode_swap_data(tx['input'])
            
            if decoded_input:
                token_in = decoded_input['tokenIn']
                token_out = decoded_input['tokenOut']
                amount_in = decoded_input['amountIn']
                
                # Calculate potential profit from sandwich attack
                profit = await self.calculate_sandwich_profit(
                    token_in,
                    token_out,
                    amount_in,
                    tx['gasPrice']
                )
                
                if profit and profit['profit_usd'] > 0:
                    return {
                        'type': 'sandwich',
                        'target_tx': tx['hash'].hex(),
                        'token_in': token_in,
                        'token_out': token_out,
                        'profit_usd': profit['profit_usd'],
                        'front_run_amount': profit['front_run_amount'],
                        'gas_cost': profit['gas_cost']
                    }
                    
        except Exception as e:
            print(f"Transaction analysis error: {e}")
            
        return None
        
    async def calculate_sandwich_profit(
        self,
        token_in: str,
        token_out: str,
        amount_in: int,
        gas_price: int
    ) -> Optional[Dict]:
        """Calculate potential profit from sandwich attack"""
        try:
            # Simulate frontrun transaction
            front_run_amount = amount_in * 2  # 2x the target transaction
            
            # Estimate gas costs
            gas_limit = 300000  # Conservative estimate
            gas_cost_wei = gas_limit * gas_price
            gas_cost_usd = self.wei_to_usd(gas_cost_wei)
            
            # Simulate price impact and backrun profit
            price_impact = await self.simulate_price_impact(
                token_in,
                token_out,
                front_run_amount
            )
            
            if price_impact:
                profit_usd = price_impact['profit'] - gas_cost_usd
                
                if profit_usd > self.min_profit * (front_run_amount / 1e18):
                    return {
                        'profit_usd': profit_usd,
                        'front_run_amount': front_run_amount,
                        'gas_cost': gas_cost_usd
                    }
                    
        except Exception as e:
            print(f"Profit calculation error: {e}")
            
        return None
        
    async def execute_sandwich(self, opportunity: Dict) -> Optional[float]:
        """Execute sandwich attack"""
        try:
            # Build frontrun transaction
            frontrun_tx = self.build_swap_tx(
                opportunity['token_in'],
                opportunity['token_out'],
                opportunity['front_run_amount'],
                priority_fee=2  # Higher priority fee to ensure frontrun
            )
            
            # Build backrun transaction
            backrun_tx = self.build_swap_tx(
                opportunity['token_out'],
                opportunity['token_in'],
                opportunity['front_run_amount'] * 1.01,  # Account for slippage
                priority_fee=1
            )
            
            # Execute transactions
            frontrun_receipt = await self.send_transaction(frontrun_tx)
            if frontrun_receipt and frontrun_receipt['status'] == 1:
                # Wait for target transaction
                await self.wait_for_transaction(opportunity['target_tx'])
                
                # Execute backrun
                backrun_receipt = await self.send_transaction(backrun_tx)
                if backrun_receipt and backrun_receipt['status'] == 1:
                    profit = opportunity['profit_usd']
                    
                    # Update wallet
                    platform_fee = profit * 0.1  # 10% platform fee
                    user_profit = profit * 0.9  # 90% to user
                    
                    self.wallet_manager.add_funds("platform", platform_fee)
                    return user_profit
                    
        except Exception as e:
            print(f"Sandwich execution error: {e}")
            
        return None
        
    async def run(self):
        """Main agent loop"""
        print("Starting MEV agent...")
        while True:
            try:
                # Monitor mempool for opportunities
                opportunities = await self.monitor_mempool()
                
                for opportunity in opportunities:
                    if opportunity['profit_usd'] > self.min_profit * self.capital:
                        profit = await self.execute_sandwich(opportunity)
                        if profit:
                            print(f"MEV profit: ${profit:.2f}")
                            
                # Brief pause between iterations
                await asyncio.sleep(1)  # 1 second delay
                
            except Exception as e:
                print(f"MEV agent error: {e}")
                await asyncio.sleep(30)  # 30 second cooldown on error
                
    def decode_swap_data(self, input_data: str) -> Optional[Dict]:
        """Decode swap transaction input data"""
        # Add implementation for decoding specific DEX swap methods
        pass
        
    def build_swap_tx(self, token_in: str, token_out: str, 
                     amount: int, priority_fee: int) -> Dict:
        """Build swap transaction"""
        # Add implementation for building DEX swap transaction
        pass
        
    async def send_transaction(self, tx: Dict) -> Optional[Dict]:
        """Send transaction to network"""
        # Add implementation for sending transaction
        pass
        
    async def wait_for_transaction(self, tx_hash: str):
        """Wait for transaction to be mined"""
        # Add implementation for transaction waiting
        pass
        
    def wei_to_usd(self, wei_amount: int) -> float:
        """Convert wei to USD"""
        # Add implementation for price conversion
        pass
        
    async def simulate_price_impact(self, token_in: str, 
                                  token_out: str, amount: int) -> Optional[Dict]:
        """Simulate price impact of swap"""
        # Add implementation for price impact simulation
        pass
