user-workspace/swarm/agents/MethodResearcherAgent.py
import asyncio
import aiohttp
import json
import time
from datetime import datetime
from bs4 import BeautifulSoup

class MethodResearcherAgent:
    def __init__(self):
        self.discovered_methods = []
        self.last_research_time = 0
        self.research_interval = 3600  # 1 hour between research sessions
        
    async def research_new_methods(self):
        """Actively research and discover new money-making methods"""
        current_time = time.time()
        if current_time - self.last_research_time < self.research_interval:
            return
        
        self.last_research_time = current_time
        
        # Sources to research from
        sources = [
            'https://www.reddit.com/r/beermoney',
            'https://www.reddit.com/r/passive_income',
            'https://www.reddit.com/r/WorkOnline',
            'https://www.reddit.com/r/freelance',
            'https://www.reddit.com/r/EntrepreneurRideAlong'
        ]
        
        async with aiohttp.ClientSession() as session:
            for source in sources:
                try:
                    # In a real implementation, you'd need proper API access or scraping capabilities
                    # This is a placeholder for demonstration
                    methods = await self._scrape_money_methods(session, source)
                    self.discovered_methods.extend(methods)
                except Exception as e:
                    print(f"Error researching from {source}: {e}")
    
    async def _scrape_money_methods(self, session, url):
        """Placeholder for actual web scraping logic"""
        # In a real implementation, this would scrape or use APIs to discover methods
        # This is a mock implementation for demonstration
        
        methods = [
            {
                'name': 'NFT_MINTING',
                'description': 'Create and sell NFT collections',
                'min_value': 50,
                'max_value': 5000,
                'difficulty': 'medium',
                'time_investment': 'medium',
                'source': url,
                'discovered_at': datetime.utcnow().isoformat()
            },
            {
                'name': 'SOCIAL_MEDIA_MANAGEMENT',
                'description': 'Manage social media accounts for businesses',
                'min_value': 500,
                'max_value': 3000,
                'difficulty': 'low',
                'time_investment': 'high',
                'source': url,
                'discovered_at': datetime.utcnow().isoformat()
            },
            {
                'name': 'CRYPTO_ARBITRAGE',
                'description': 'Profit from cryptocurrency price differences',
                'min_value': 100,
                'max_value': 10000,
                'difficulty': 'high',
                'time_investment': 'high',
                'source': url,
                'discovered_at': datetime.utcnow().isoformat()
            }
        ]
        
        return methods
    
    async def validate_methods(self):
        """Validate discovered methods for viability"""
        validated_methods = []
        
        for method in self.discovered_methods:
            # Real validation would involve checking:
            # - Market demand
            # - Legal compliance
            # - Technical feasibility
            # - Profit margins
            
            # Mock validation for demonstration
            if method['difficulty'] != 'high' and method['max_value'] > 100:
                validated_methods.append(method)
        
        self.discovered_methods = validated_methods
    
    async def analyze_market_trends(self):
        """Analyze current market trends for money-making opportunities"""
        trends = {
            'AI_CONTENT': {'growth': 0.25, 'demand': 'high', 'competition': 'medium'},
            'NFT_TRADING': {'growth': -0.1, 'demand': 'medium', 'competition': 'high'},
            'DROPSHIPPING': {'growth': 0.05, 'demand': 'high', 'competition': 'high'},
            'FREELANCE_WRITING': {'growth': 0.15, 'demand': 'medium', 'competition': 'medium'},
        }
        
        # Sort trends by potential
        sorted_trends = sorted(
            trends.items(),
            key=lambda x: x[1]['growth'] * (1 - (0.3 if x[1]['competition'] == 'high' else 0.1)),
            reverse=True
        )
        
        return sorted_trends[:5]
    
    def get_recommendations(self):
        """Get top recommended methods based on research"""
        recommendations = []
        
        for method in self.discovered_methods:
            recommendation = {
                'method_name': method['name'],
                'potential_earnings': (method['min_value'] + method['max_value']) / 2,
                'difficulty': method['difficulty'],
                'time_investment': method['time_investment'],
                'recommendation_score': self._calculate_recommendation_score(method)
            }
            recommendations.append(recommendation)
        
        return sorted(recommendations, key=lambda x: x['recommendation_score'], reverse=True)[:10]
    
    def _calculate_recommendation_score(self, method):
        """Calculate a method's recommendation score"""
        earnings_weight = 0.4
        difficulty_weight = 0.3
        time_weight = 0.3
        
        # Simple scoring system
        earnings_score = min(1.0, method['max_value'] / 1000)
        difficulty_score = 1.0 - (0.0 if method['difficulty'] == 'low' else 0.5 if method['difficulty'] == 'medium' else 1.0)
        time_score = 1.0 - (0.0 if method['time_investment'] == 'low' else 0.5 if method['time_investment'] == 'medium' else 1.0)
        
        return (earnings_score * earnings_weight + 
                difficulty_score * difficulty_weight + 
                time_score * time_weight)